+++ Krowi_ExtendedVendorUI.lua
@@ -9,7 +9,6 @@ addon.IsDragonflightRetail = major == "10";
 
 -- [[ Ace ]] --
 addon.L = LibStub(addon.Libs.AceLocale):GetLocale(addonName);
--- addon.InjectOptions:SetLocalization(addon.L);
 
 -- [[ Load addon ]] --
 local loadHelper = CreateFrame("Frame");
@@ -17,7 +16,7 @@ loadHelper:RegisterEvent("ADDON_LOADED");
 
 function loadHelper:OnEvent(event, arg1, arg2)
     if event == "ADDON_LOADED" then
-        if arg1 == "Krowi_MerchantFrameExtended" then -- This always needs to load
+        if arg1 == addonName then -- This always needs to load
             addon.Options:Load();
 
             addon.Gui.MerchantItemsContainer:LoadMaxNumItemSlots();