+++ Localization/enUS.lua
@@ -11,7 +11,7 @@ L["TAB"] = tab;
 -- Load strings into the localization that are already localized by Blizzard
 addon.GlobalStrings.Load(L);
 
--- [[ https://legacy.curseforge.com/wow/addons/krowi-merchant-frame-extended/localization ]] --
+-- [[ https://legacy.curseforge.com/wow/addons/krowi-extended-vendor-ui/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
 
 L["Are you sure you want to hide the options button?"] = [=[Are you sure you want to hide the options button?
@@ -27,6 +27,7 @@ L["Default value"] = true
 L["Discord"] = true
 L["Discord Desc"] = "Open a popup dialog with a link to the {serverName} Discord server. Here you can post comments, reports, remarks, ideas or anything else related."
 L["Hide"] = true
+L["Icon Left click"] = "for quick layout options."
 L["Icon Right click"] = "for Options."
 L["Options button"] = true
 L["Options Desc"] = "Open the options that are also available from the options button in the merchant window."
@@ -42,4 +43,3 @@ L["Wago"] = true
 L["Wago Desc"] = "Open a popup dialog with a link to the {addonName} {wago} page."
 L["WoWInterface"] = true
 L["WoWInterface Desc"] = "Open a popup dialog with a link to the {addonName} {woWInterface} page."
-